requirements_hydra.txt
hydra-core>=1.3.0
omegaconf>=2.3.0
pytorch-lightning>=2.0.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
torchmetrics>=1.0.0
pillow>=9.0.0
numpy>=1.21.0
matplotlib>=3.5.0
tqdm>=4.64.0
scikit-learn>=1.1.0
seaborn>=0.11.0
pandas>=1.5.0 